Moved out helper functions away from factories file into testUtilities one
